Fix imports and code to align with newly split types
