+++ javascript/packages/orchestrator/src/paras-decorators/acala.ts
@@ -1,4 +1,4 @@
-import { Node } from "../types";
+import { Node } from "../sharedTypes";
 
 // Acala genesis node key type
 export type GenesisNodeKey = [string, string, { [key: string]: string }];
