+++ javascript/packages/orchestrator/src/paras-decorators/bifrost.ts
@@ -1,4 +1,4 @@
-import { Node } from "../types";
+import { Node } from "../sharedTypes";
 
 // Bifrost genesis node key type
 export type GenesisNodeKey = [string, string, { [key: string]: string }];
