+++ javascript/packages/orchestrator/src/paras-decorators/equilibrium.ts
@@ -4,7 +4,7 @@ import {
   readAndParseChainSpec,
   writeChainSpec,
 } from "../chainSpec";
-import { Node } from "../types";
+import { Node } from "../sharedTypes";
 
 // Acala genesis node key type
 export type GenesisNodeKey = [string, string, { [key: string]: string }];
