+++ javascript/packages/orchestrator/src/providers/k8s/resources/bootNodeResource.ts
@@ -1,4 +1,4 @@
-import { Node, ZombieRole } from "../../../types";
+import { Node, ZombieRole } from "../../../sharedTypes";
 import { NodeResource } from "./nodeResource";
 import { Container, PodSpec, Volume } from "./types";
