+++ javascript/packages/orchestrator/src/providers/k8s/resources/types.ts
@@ -1,4 +1,4 @@
-import { envVars, Resources, ZombieRoleLabel } from "../../../types";
+import { envVars, Resources, ZombieRoleLabel } from "../../../sharedTypes";
 
 type ContainerResource = Resources["resources"];
