+++ javascript/packages/orchestrator/src/providers/native/substrateCliArgsHelper.ts
@@ -1,4 +1,4 @@
-import { SubstrateCliArgsVersion } from "../../types";
+import { SubstrateCliArgsVersion } from "../../sharedTypes";
 import { getClient } from "../client";
 
 export const getCliArgsVersion = async (
