+++ javascript/packages/orchestrator/src/spawner.ts
@@ -19,13 +19,8 @@ import { Network, Scope } from "./network";
 import { NetworkNode } from "./networkNode";
 import { getProvider } from "./providers";
 import { Client } from "./providers/client";
-import {
-  Node,
-  NodeMultiAddress,
-  Parachain,
-  ZombieRole,
-  fileMap,
-} from "./types";
+import { NodeMultiAddress, fileMap } from "./types";
+import { Node, ZombieRole, Parachain } from "./sharedTypes";
 const debug = require("debug")("zombie::spawner");
 
 export const spawnNode = async (
